refactor(server): extract MCP tool call tracker to middleware/protocol_injection.py
Task E of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history
(_mcp_tool_call_tracker, ~171 lines). Removed the Task D import block
from server.py (state now used only by the tracker, which owns it
from its new home) and the now-dead json/Request/JSONResponse/
compaction/documentation/helpers imports. Middleware registration
order in server.py is unchanged: tracker, then auth, then rate_limit
(3 app.middleware("http") calls, only the import source changed).

Two test files reach into the tracker's module-level state by
qualified name (test_protocol_lite.py mutates in place via
server._protocol_delivered_sessions.clear() etc.; test_http_tools.py
rebinds via server._protocol_delivered_sessions = set()). A verbatim
move cannot keep these literally unchanged: server.py no longer owns
these names, and re-exporting them would recreate the exact
two-names-one-object footgun that caused the original break (a rebind
via one name doesn't reach a cached name-import elsewhere). Repointed
both files' handles to their new homes (core.protocol_delivery_state
for the state dicts/prune function, middleware.protocol_injection for
the tracker) via importlib.import_module, matching the isolation
harness's existing pattern. No assertion or test logic changed.

Verified before repointing that this actually works, not just that
tests pass: _protocol_delivered_sessions is never name-imported into
protocol_injection.py (the tracker only reaches it via function calls
defined in protocol_delivery_state.py, which do live global lookups in
their own module dict at call time) -- so a rebind via
protocol_delivery_state._protocol_delivered_sessions correctly
propagates. Note: _protocol_call_counts/_protocol_delivery_lock/
_PROTOCOL_LITE_INTERVAL ARE name-imported into protocol_injection.py
and dereferenced directly in the tracker body; this remains a latent
footgun if anything ever rebinds (rather than mutates in place) those
specific names, but no current test exercises that path, and fixing it
is out of scope for a pure wiring-only extraction.

Removed 3 now-genuinely-unused `server = load_isolated_server(...)`
bindings in the two test files (ruff F841) where server was only ever
used for the now-repointed tracker/state access; the call itself is
kept for its side effects (module purge/reimport, temp DB env setup).

Functional check boots the real FastAPI app, confirms auth middleware
still gates /mcp before the tracker can run (401 on bad key -- proves
registration order intact), and the full targeted suite (62 tests
across test_http_tools.py, test_protocol_lite.py,
test_http_security_regressions.py, test_http_rate_limit.py,
test_http_app.py) passes with real assertions against real injected
protocol/compaction content.
